Add command line options for the json history
Let the user control the json history by adjusting interval and size.
Change the json history to be allocated on the heap instead of the stack
as it's no longer static in size.
Also change Webview history loading timeout to scale with the history
size so it does not time out when using large history sizes.