Improve Webview performance (altitude changes)
The color of the plane track is changed according to its altitude.
For every change a new OpenLayer Linestring object is created, even when
the airplane is flying basically level and only the readout changes by
25 ft.
Improve performance by only changing color if the altitude changed by
more than 200 feet compared to the color being displayed.
According to my eyes even 500 feet difference does not create a visible
discontinuity in the color of the plane track.