Webview: Fix position comparison and wrong append
The position variable is an array and needs to be compared element-wise.
Therefore change position comparison to an element-wise comparison.
Due to this comparison not working the else branch of if(est_track) was
never reached in the program flow.
With the estimated part just starting the if branch of if(est_track) will
append projHere. The else branch then appends it again as projPrev.
Changing back to solid line a gap is created between the dotted line and
the new solid line because projPrev is only appened to the solid line.
Changing projPrev to projHere in the else branch fixes both problems.