Webview: Improve estimated track handling
Intermittent positions are displayed as estimated (dotted) track. The
handling of that feature was confusing and differed significantly
between history and live data.
Clean it up so it works as intended.
Check if it was already stale at the time of the data set creation,
not only the time difference to the last position as this is
insufficient when data sets are created in larger intervals. This is
still not sufficient to detect every 5 second gap in position data but
it's an improvement.
Also call updateTick with both arguments so they can be used when
calling updateTrack in updateTick.