Webview: fix append handling
The "tail" that is the latest fixed segment is set to the current
position. Therefore tail_update needs to be set to the current time not
head_update. head_update is updated whenever a new position is received
but at the time of the assignment has not been yet assigned the new
value.
To compensate for the somewhat reduced frequency of positions being
saved set the minimum time after which a new position is save to 4
seconds. Times are not saved as whole seconds so use > instead of >=.