Webview: replace head_update with prev_position_time
head_update saves the timestamp of the plane's current position in the
linestring segments which makes no sense because the position is
not always part of the linestring. Save it as part of the plane object
and call it prev_position_time because it's used as the timestamp of the
previous position of the airplane.