Webview: replace tail_update with append_time
Whenever a point is appended to the current line segment, tail_update is
updated. There is no reason this should stick around after moving
on to the next segment, it is only used to decide if the current point
is to be appended. For consistency add an append_time update when drawing
estimated segments even if it's not needed.