Webview: avoid appending existing points
Check if the point is already part of the line segment before appending
it. Checking only longitude is enough because this check is not
critical.
Also fixes history_size accounting which only added 2 points before. My
understanding of the variable is the counting of the points saved in
line segments.