Change ro-interval default to something more reasonable
While every MLAT system worthwhile is independent of when the message is
transferred to the feed client, the UTC timestamp give to each position
or altitude depends on the system clock or a clock managed by the feed
client itself.
Having up to 1 second delay when the message count is low is
unnecessary, even slow computers or connections should be able to handle
10 TCP packets per second.
It is of course up to the user to configure it, but this value seems
like a more reasonable default.