correctly set position_valid when doing global CPR
Using combine_validity which uses the older cpr_valid.update time to
determine the position_valid.update time is wrong. The update time of
the position is the reception time of the current CPR.
The position_valid.update time is also used for the speed check which
relies on the actual time of the position.